refactor: rename file and fix typos
