+++ packages/mcp-auth/src/utils/create-verify-jwt.ts
